persist: remove error output from conditional_seal
We had previously changed the conditional seal implementation to not
produce errors and instead retry if there was an error. This commit
removes the error output stream that was statically guaranteed to be empty
(as it came from a Result stream that used ok_err, but the result
stream was only ever populated with Oks).