persist,dataflow: make source rendering deterministic on workers
Before, when determining the common upper seal timestamp for involved
streams, we would make fallible calls into persistence. Based on the
result, we would either render one dataflow graph or a graph with a
different shape. This does not work with timely dataflow and leads to
unexpected results.

Now, we determine the seal timestamps of involved streams in the
coordinator and send them along in the persistence description.

Fixes #9509