Fix lowering panic with LATERAL references in ROWS FROM
After #9623 a new scope is always defined for the RHS of any join,
LATERAL or not. If a relation is pushed down to the RHS of a join after
name resolution has already happened we must update the column references
leaving the scope of that relation to catch up with the scope of the
join. That is what happens when creating the FULL OUTER used for
supporting ROWS FROM.