sql: correct handling of col refs in ORDER BY/DISTINCT ON
We weren't quite PostgreSQL compatible with our handling of column
references in ORDER BY and DISTINCT ON clauses. This commit corrects
that behavior as described within, and adds a section to the `SELECT`
documentation explaining the mess.

Correcting this issue means we can also get rid of the horrible
`priority` hack we used to carry around in the scope.