persist: make seal() operator communicate when it's not done
Previously, the seal operator could be destroyed by the Timely scheduler at any
point when it deemed that that the operator would not receive any more messages.

This made testing behaviors around closing operators more complex. This commit
changes the behavior to instead indicate to timely that the operator cannot be
destroyed while it still has pending futures. It is still the operator's responsibility
to properly clear all pending futures if it wants to exit immediately. This commit
also lets us simplify the regression test for #9419.