kafka_sink: Empty write frontier with empty input frontier (#9493)
When the input frontier is empty and there are no pending rows, we should clear the write frontier signaling that there will be no more writes to the sink.

Also include all dataflow workers in the RenderState (with an empty frontier for all non-active writers) so that frontier changes are computed correctly.

Motivation
When using an exactly-once kafka sink that is sourced by a non-tailed File (critically: a source that moves into the SourceState::Done), we created an unbounded number of timestamp bindings as the sink write frontier never advances. Because of this, we're not able to compact timestamp bindings. This is less an issue than it used to be after #9379 but we still ought to fix it.