testdrive: Allow progress in verify-timestamp-compaction (#9716)
Under heavy load with multiple workers, it's possible to not compact down to just one or two timestamp bindings. However, this doesn't mean we haven't made progress. Therefore, add a new option to the verify-timestamp-bindings command in testdrive to allow for compaction to have proceeded past the highest timestamp binding of the first examination of the catalog.

This showed up in the nightly build tests which run with 32 workers. The test consistently failed even though compaction progress was in fact being made