materialized: perform log filtering on the fmt layers
Extracted from #9665.

The idea here is to perform the filtering specified by the
`--log-filter` argument at the last possible moment, right before the
message is emitted to disk. This ensures that other log messages (e.g.,
trace and debug logs) are available to other parts of the logging
subsystem that may want to do something smart with them.

This already pays dividends in that the mz_log_message_total metric can
keep track of the log messages that don't get written out to the file.
It also paves the way for integrations with e.g. the Tokio console (see

Co-authored-by: Gus Wynn <gus@materialize.com>