xcompile: use first-party cross-compiling toolchains on macOS
We're now building our own cross-compiling toolchains for macOS in
MaterializeInc/homebrew-crosstools. Switch xcompile over to using those.
This is a bit safer, as we're no longer downloading random binaries from
the internet, and also lets us control our own destiny, e.g., if we ever
want to update the version of binutils/GCC/glibc/Linux in the toolchain.