ci: add instructions for spinning up CI-like infrastructure
For security reasons, it's hard to give all engineers access to running
CI agents. But we can do something almost as good, which is to give all
engineers the ability to spin up an EC2 instance that looks nearly
identical to an actual CI agent.

This commit adds instructions for doing that. It also makes some tweaks
to the `bin/scratch` infrastructure to support AMIs whose default name
is "ec2-user" rather than "ubuntu".