demo/chbench: remove old benchmarking cruft
This benchmarking infrastructure hasn't been used in a while, and has
been replaced by the new "cloudbench" suite of tools. In particular,
it's tangled up with the chbench mzcompose.yml, which was meant to be a
user-facing demo rather than the basis of serious benchmarks, and has
started to break the actual demo (#9670).

Fix #9670.
Touches #9773.