mzcompose: improve error message output
Introduce a new UIError exception type that can nicely format an error
message along with a hint, and use it everywhere. This new exception
type subsumes the existing exception hierarchy in the `errors` module;
there's no point in having a complicated exception hierarchy when they
all ultimately result in the same error message being printed to the
console.