test: add dedicated kafka upsert benchmark
This commit adds a dedicated kafka upsert benchmark using the new mzworkflows
framework but not within the feature-benchmarks framework.

This benchmark is an open loop benchmark that takes as its main arguments
- a desired QPS rate for messages to send to Kafka
- the number of seconds to write for.

It's open loop in the sense that the rate of message generation / insertion
to Kafka is decoupled as much as possible from the rate at which messages are
ingested into Materialize. We intentionally don't wait for Materialize to catch
up to the last set of messages before sending the next set of messages.

The benchmark allows for flexibility in the following other dimensions:
- key cardinality
- value size
- whether or not persistence is enabled

Given the following inputs the benchmark sends approximately `records_per_second`
messages to Kafka per second and corrects for delays caused by slow input generation
or slow querying. It also queries Materialize and tracks how many records Materialize
has ingested, and reports the lag relative to the number of messages already inserted.

The benchmark currently doesn't support:
- data formats other than bytes
- specifying the number of partitions (currently hardcoded to 4)
- specifying the number of materialize workers