sql: support non-nested arbitrary table functions
Teach the planner how to plan arbitrary table functions in SELECT lists,
and remove the corresponding ast rewrites that did "only one unique
table func" version of this feature. Implement this using ROWS FROM
planning. An additional ordinality flag is added to enable keeping each
table func's ordinality column instead of the single coalesced ordinality
column. This column is used to tell the Scope whether the entire row is
NULL if requested in * expansion.

The only remaining unsupported feature for table functions is table
functions as arguments for other table functions, so not yet closing
out the issue.

See #1546