doc/user: pin demos to latest stable release (#9830)
The current demo instructions have users run demos from the latest tip
of main. This is dangerous because those demos are running against an
unstable and possibly broken version of Materialize. Worse, the Docker
images for the latest tip of main take a few minutes to become
available; if you get unlucky, you'll accidentally wind up trying to
compile the latest tip from source, which most users running demos are
not prepared to do.

This commit updates the instructions so that users get a clone of the
demo from the latest stable release. It also adds `--depth=1` to the
`git clone` call, which makes the clone *much* faster because users
don't need to check out the repository in its entirety.