dataflow,sources: only restore timestamp bindings on workers that need them
Before, we would restore all existing bindings to all workers, meaning
that workers that are not responsible for a given source id/partition id
combination nevertheless had bindings stored for them. This is
problematic because these workers periodically report those bindings,
meaning they potentially send old bindings to the coordinator that have
been superseded.

This lead to the situation that the timestamp bindings we store in
SQLite have bindings for lower offsets after bindings with higher
offsets, when sorted by the time at which the binding was reported to
the coordinator. This in turn, would lead to panics when restoring
timestamp bindings, because we check that offsets don't go backwards.

The problem only surfaces when restarting materialized multiple times,
which is what some of the newer persistence tests do. Although the
problem is not related to the newly added source persistence code (which
also deals with timestamp bindings).

Fixes #9777