Get rid of redundant `TailResponse::Complete` variant.
In Timely, a dataflow operator (assuming it has retained no
capabilities) is "complete" if and only if its input frontier is
empty. Thus, there is no reason for a separate complete message
to exist, since we can communicate the same information by sending
a progress message with an empty frontier.

This causes no visible changes, because the coordinator
was already cleaning up the tail anyway in response to an empty
frontier, and therefore it was impossible for it
to ever receieve a `Complete` notification.