mzcompose: make workflow steps normal Python methods
Make the WorkflowStep subclasses normal Python methods on the Workflow
class. Removing the magic will make it possible to typecheck the
mzworkflows.py files to catch errors.

This commit is entirely code movement. I was very careful not to change
the behavior of any of the steps. At most I deleted unused parameters
and steps.

See #9809.