mzcompose: merge Workflow into Composition
Merge the Workflow object into the Composition object in mzcompose.
Workflow functions now receive a `c: Composition` as their first
argument rather than a `w: Workflow`.

This is a minor change but an important one, I think. Previously the
conceptual model was "a workflow is a function that receives a workflow
as an input". That's weird! In the new model, the workflow *is* the
function, and it gets provided a composition to manipulate. This is
much cleaner, I think.

It also means you can't accidentally pass the wrong workflow object to
another workflow. In the previous API it was possible to write code like
this:

    def workflow_one(w: Workflow):
        workflow_two(w)

    def workflow_two(w: Workflow):
        print(w.name)

Here, `./mzcompose run two` would print "two" but `./mzcompose run one`
would print "one". The new model avoids this:

    def workflow_one(c: Composition):
        workflow_two(c)

    def workflow_two(c: Composition):
        # No incorrect workflow name to refer to!

Argument parsing is also simplified. Workflows now optionally receive a
premade WorkflowArgumentParser as input:

    def workflow_args(c: Composition, parser: WorkflowArgumentParser) -> None:
        parser.add_argument("--foo")
        args = parser.parse_args()
        ...

This is a bit smoother in the simple case. When one workflow wants to
call another workflow with arguments, it's still easy. You can't call
the workflow directly, but you can invoke `c.workflow`:

    def workflow_specific(c: Composition) -> None:
       c.workflow("general", ["--foo"])

    def workflow_general(c: Composition, parser: WorkflowArgumentParser) -> None:
        parser.add_argument("--foo")
        args = parser.parse_args()
        ...

There are no known behavior changes in this commit.