persist: pass around Metrics wrapped in an Arc
This is almost entirely forward looking. While it's safe for Metrics to
directly derive Clone and for us to clone it in the few places we do
now, it's a big struct and these Clones are expensive. This doesn't
matter now, but at some point we'll need to pass a copy into e.g. a
snapshot so if it hits a transient error, we can try the retries.