sql: use cast target type as invented column name fallback
To match PostgreSQL, use cast's target type as an invented column name
if the inner expression doesn't have a different name to use. This will
be tested in the next commit, which imports a bunch of tests from
PostgreSQL that exercise this feature.

Since I'm in the area, also correct one other instance of misnamed
columns that I spotted: bool and interval literals are named `bool` and
`interval`, respectively. This (correctly) breaks some upgrade tests,
which are adjusted here to avoid testing the behavior of column name
inference.

Fix #8010.