dataflow,sources: fix logic that filters restored (persisted) timstamp bindings
Before, get_binding() would return a "binding" based on the Proposer,
even when there are no real stored bindings. This could lead to the case
were we thought that there we already bindings that supersede the
bindings restored from persistence but there were not, and we would
therefore end up in a situation without any bindings, meaning the Kafka
source would re-read all its input from the beginning again.

Fixes #9687