dataflow,sources: only restore (persistence) timestamp bindings on workers that need them
This is the same as #9777, but for bindings stored in persistence
instead of the bindings stored in SQLite/coord.

Before, we would restore all existing bindings on all workers, meaning
that workers that are not responsible for a given source id/partition id
combination nevertheless had bindings stored for them. This is
problematic because these workers periodically report those bindings,
meaning they potentially send old bindings to the coordinator that have
been superseded.

This lead to the situation that the timestamp bindings we store in
SQLite have bindings for lower offsets after bindings with higher
offsets, when sorted by the time at which the binding was reported to
the coordinator. This in turn, would lead to panics when restoring
timestamp bindings, because we check that offsets don't go backwards.

Additionally, we need to make sure that retractions are only emitted
from the worker that is responsible for the partition of the binding.