issue_template: only destroy load test instances
The release template has a command to destroy all load test instances which actually
destroys the last two instances spawned by that user.

Change the command provided to first filter for load test instances for the current
two load tests. This is still a brittle and best effort solution as we may add
more load tests in the future, or someone might run another instance with the
name 'chbench'.