[kafka sink] Remove blocking tx calls from dataflow thread (#10026)
Take 2 of #9909 (including small fix from #9982).

The v1 implementation held a borrow to the frontiers through each invocation of sink which made it difficult for `shared_frontiers.try_borrow_mut` to obtain the borrow and advance the frontier.  Because the frontier was not easily advanced, the kafka sink never learned of any `closed_ts`s and, therefore, never wrote any records out to the sink.

Clone the frontier before passing it into the `$body` of the `async_op` macro.

### Motivation
Fixes #6035 

### Tips for reviewer
- c95fef3f2c168f9d5784c657e5b0d4ed23068c8d is the two original PRs merged
- 2c715a497abc5d92ad67e985dbb17f18b82d5f20 is the new code.

### Checklist
- [x] This PR has adequate test coverage / QA involvement has been duly considered.
I ran `kafka-exactly-once-sink.td` many times in CI and manually checked that it does not have the performance regression of the original #9909.

Additionally, I found a minimal repro (aka a repro that didn't involve the sink kafka tx calls actually being async) and verified this change fixed it.