[testdrive] Allow for missing initial items in kafka-verify (#10093)
Add a new greedy-search option for the kafka-verify testdrive command. This allows us the search for a sequence of records that does not begin at the beginning of the topic.

### Motivation
Will help unblock #9514 . There's no reason that we need to start with a BEGIN record and, as shown in this issue, it can cause us some trouble. I've pulled this out because the direct fix to that issue is in a complex PR (that I'll link after I create shortly)