testdrive: invoke real protoc in protobuf-compile action
Extracted from #10079. See that PR for justification on the migration of
rust-protobuf.

This commit changes testdrive to invoke the proper `protoc` binary
that's built by the protobuf-src crate. Using a real Protobuf compiler
is more representative of how our end users build descriptors.

The blocker to this historically has been that we didn't want protobuf
to become a build dependency via `brew install protobuf`, since
different developers wind up with different versions of protobuf
installed. The protobuf-src crate fixes that by vendoring a particular
version of the protobuf source code and building it from source.