core,test: Allow stack traces on SIGSEGV/SIGBUS to be disabled
Attempting to dump stack traces in case of SIGSEGV/SIGBUS may
cause the core file to become garbled and contain the strack frames
of the stack dumper and not the original stack frames that lead to the
crash.

Provide an option to disable stack traces on SIGSEGV/SIGBUS and
use it to obtain core files in the Antithesis environment
in order to debug one particular issue that only happens there.