feature-benchmark: Sleep 10ms when retrying a query
Previously, testdrive retried each query until it succeeds, without
any delay. Introduce a 10ms delay in order to burn less CPU in
busy waiting.

The noise floor when it comes to operations that require retries,
such as waiting for a source to become fully hydrated is such that
precise measurements with resolution less than 10ms are unlikely
in practice.