dataflow-types: make AwsConfig::load Result-less too
Now that constructing AWS clients is Result-less, it's worth trying a
little bit harder to make AwsConfig::load Result-less too. The reason I
didn't do this previously is because serializing/deserializing an
http::Uri requires a newtype and a special crate, but now it seems well
worth it for the net reduction in code.