[kafka sink] Rewrite sink to take advantage of async_op (#10044)
Remove the explicit state machine-based main logic loop for the kafka sink. It's no longer necessary now that we have an async operation framework. Note that most of the complexity of this rewritten version lies in the error handling. That is not in a final state but I think it's an incremental improvement over what we had previously.

I've additionally made a change to use flush rather than needing to rely on checking in_flight_count on the kafka producer.

At the end of an invocation of the body of the async_op, we should now have the following invariants:

All rows that have occurred at a time that is less than both 1) the input frontier and 2) the durability frontier of the inputs, have been send to kafka
The kafka producer has been flushed
All other rows received are maintained in pending_rows until time advances sufficiently that they are able to be send
The existing tests should be sufficient. I'll also do some brief perf testing before merging to make sure there are no regressions

Motivation
This makes the code more clear and will allow us to do some error-handling refactoring more easily in the future.

Fixes #9980