testdrive: switch to anyhow
Testdrive predates anyhow, and so had its own homegrown error type that
was mostly equivalent in functionality to `anyhow::Error`. But for
historical reasons, the `testdrive::Error` type was not used throughout
the entire crate, and so much of testdrive dealt in `String` errors.
This made writing testdrive actions unnecessarily verbose; lots of calls
like:

    fallible_op().map_err(|e| format!("fallible op failed: {:#}", e)

This commit switches testdrive to use `anyhow::Error` throughout the
crate. Much of the error handling goop becomes vastly less verbose, and
even when it doesn't, it becomes much more familiar.

The only casualty is the loss of error "hints". These weren't that
helpful, and in most cases have been largely mitigated by adding a bit more
context to the error message itself.

Fix #10172.