mzbuild: always perform docker pull on acquire()
perform 'docker pull' when acquiring an image even if self.publish
is set to False. This way a fresh copy of images such as
'materialized:latest' will always be fetched from the Docker hub.

Prior to this fix, mzcompose would use an old image if one was
cached locally, defeating the purpose of performing tests against
'latest'.