dataflow: compact persistent sources on AllowSourceCompaction
The compaction since of persisted sources is upper-bounded by two
frontiers: a) the compaction that the coordinator allows, and b) the
seal frontier of the involved persistent collections.

We need to obey a) because that's what the coordinator wants, and b)
because if we persisted past (or up to) the seal frontier we would not
be able to distinguish between valid and invalid updates based on the
upper seal frontier when restarting.

kafka-ingest-open-loop benchmark results on my machine (Ryzen 5950x,
64gb ram, 2TB SN850 SSD):

this change (with compaction):
$ MZ_WORKERS=4 bin/mzcompose --find kafka-ingest-open-loop run default --num-seconds=360 --records-per-second=20000 --upsert --num-keys 1000 --enable-persistence
Finished after 367.098s sent and ingested 7200000 records. max observed lag 109568 records.
size of persist directory: 604K

main:
$ MZ_WORKERS=4 bin/mzcompose --find kafka-ingest-open-loop run default --num-seconds=360 --records-per-second=20000 --upsert --num-keys 1000 --enable-persistence
Finished after 364.587s sent and ingested 7200000 records. max observed lag 109608 records.
size of persist directory: 182M