coord,persist: fetch latest persist frontiers when rendering source
Before, we were fetching the latest seal/compaction frontiers when
starting up. We did this, because the dataflow rendering paths where not
fallible.

Recently, these rendering paths gained the ability to return errors, so
now we can call into persistence right when rendering the source. This
fixes a problem where previously creating a materialized view, dropping
it, and then re-creating it would result in errors because the rendering
code would still assume the frontiers as they were at the time the
coordinator had started up.

The problem would manifest as error messages in the log, and the
rendered view would be wedged with a persistence-related error message.
I added a test that verifies that we can drop and re-render a source
successfully.

This does change where we keep "persist details". Before, we were
keeping them in the `SourceConnector`, but these were "resolved"
details, that is with the frontiers as they were when the coordinator
booted up. Now we keep the serialized (or "unresolved") form of the
persist details in the catalog entry `Source` and resolve them when
rendering the source to get the latest frontiers.